// apps/onestock-extensions/extensions/delivery-promise-function/generated/api.ts
export type Maybe<T> = T | null;
// ...
export type Attribute = {
  __typename?: 'Attribute';
  /** The key or name of the attribute. For example, `"customer_first_order"`. */
  key: Scalars['String']['output'];
  /** The value of the attribute. For example, `"true"`. */
  value?: Maybe<Scalars['String']['output']>;
};
// ...
export type Cart = HasMetafields & {
  __typename?: 'Cart';
  /**
   * The custom attributes associated with a cart to store additional information. Cart attributes
   * allow you to collect specific information from customers on the **Cart** page, such as order notes,
   * gift wrapping requests, or custom product details. Attributes are stored as key-value pairs.
   */
  attribute?: Maybe<Attribute>;
  /**
   * Information about the customer that's interacting with the cart. It includes details such as the
   * customer's email and phone number, and the total amount of money the customer has spent in the store.
   * This information helps personalize the checkout experience and ensures that accurate pricing and delivery options
   * are displayed to customers.
   */
  buyerIdentity?: Maybe<BuyerIdentity>;
  /**
   * A breakdown of the costs that the customer will pay at checkout. It includes the total amount,
   * the subtotal before taxes and duties, the tax amount, and duty charges.
   */
  cost: CartCost;
  /** The items in a cart that are eligible for fulfillment and can be delivered to the customer. */
  deliverableLines: Array<DeliverableCartLine>;
  /**
   * A collection of items that are grouped by shared delivery characteristics. Delivery groups streamline
   * fulfillment by organizing items that can be shipped together, based on the customer's
   * shipping address. For example, if a customer orders a t-shirt and a pair of shoes that can be shipped
   * together, then the items are included in the same delivery group.
   *
   * In the [Order Discount](https://shopify.dev/docs/api/functions/reference/order-discounts) and
   * [Product Discount](https://shopify.dev/docs/api/functions/reference/product-discounts) legacy APIs,
   * the `cart.deliveryGroups` input is always an empty array. This means you can't access delivery groups when
   * creating Order Discount or Product Discount Functions. If you need to apply discounts to shipping costs,
   * then use the [Discount Function API](https://shopify.dev/docs/api/functions/reference/discount)
   * instead.
   */
  deliveryGroups: Array<CartDeliveryGroup>;
  /**
   * The items in a cart that the customer intends to purchase. A cart line is an entry in the
   * customer's cart that represents a single unit of a product variant. For example, if a customer adds two
   * different sizes of the same t-shirt to their cart, then each size is represented as a separate cart line.
   */
  lines: Array<CartLine>;
  /**
   * The additional fields on the **Cart** page that are required for international orders in specific countries,
   * such as customs information or tax identification numbers.
   */
  localizedFields: Array<LocalizedField>;
  /**
   * A [custom field](https://shopify.dev/docs/apps/build/custom-data) that stores additional information
   * about a Shopify resource, such as products, orders, and
   * [many more](https://shopify.dev/docs/api/admin-graphql/latest/enums/MetafieldOwnerType).
   * Using [metafields with Shopify Functions](https://shopify.dev/docs/apps/build/functions/input-output/metafields-for-input-queries)
   * enables you to customize the checkout experience.
   */
  metafield?: Maybe<Metafield>;
  /** The physical location where a retail order is created or completed. */
  retailLocation?: Maybe<Location>;
};
// ...
export type CartDeliveryGroup = {
  __typename?: 'CartDeliveryGroup';
  /**
   * Information about items in a cart that a customer intends to purchase. A cart line is an entry in the
   * customer's cart that represents a single unit of a product variant. For example, if a customer adds two
   * different sizes of the same t-shirt to their cart, then each size is represented as a separate cart line.
   */
  cartLines: Array<CartLine>;
  /** The shipping or destination address associated with the delivery group. */
  deliveryAddress?: Maybe<MailingAddress>;
  /**
   * The delivery options available for the delivery group. Delivery options are the different ways that customers
   * can choose to have their orders shipped. Examples include express shipping or standard shipping.
   */
  deliveryOptions: Array<CartDeliveryOption>;
  /** The type of merchandise in the delivery group. */
  groupType: CartDeliveryGroupType;
  /**
   * A [globally-unique ID](https://shopify.dev/docs/api/usage/gids)
   * for the delivery group.
   */
  id: Scalars['ID']['output'];
  /** Information about the delivery option that the customer has selected. */
  selectedDeliveryOption?: Maybe<CartDeliveryOption>;
};
// ...
export type CartDeliveryOption = {
  __typename?: 'CartDeliveryOption';
  /**
   * A unique identifier that represents the delivery option offered to customers.
   * For example, `Canada Post Expedited`.
   */
  code?: Maybe<Scalars['String']['output']>;
  /** The amount that the customer pays if they select the delivery option. */
  cost: MoneyV2;
  /**
   * The delivery method associated with the delivery option. A delivery method is a way that merchants can
   * fulfill orders from their online stores. Delivery methods include shipping to an address,
   * [local pickup](https://help.shopify.com/manual/fulfillment/setup/delivery-methods/pickup-in-store),
   * and shipping to a [pickup point](https://help.shopify.com/manual/fulfillment/shopify-shipping/pickup-points),
   * all of which are natively supported by Shopify checkout.
   */
  deliveryMethodType: DeliveryMethod;
  /** A single-line description of the delivery option, with HTML tags removed. */
  description?: Maybe<Scalars['String']['output']>;
  /**
   * A unique, human-readable identifier of the delivery option's title.
   * A handle can contain letters, hyphens (`-`), and numbers, but not spaces.
   * For example, `standard-shipping`.
   */
  handle: Scalars['Handle']['output'];
  /**
   * The name of the delivery option that displays to customers. The title is used to construct the delivery
   * option's handle. For example, if a delivery option is titled "Standard Shipping", then the handle is
   * `standard-shipping`.
   */
  title?: Maybe<Scalars['String']['output']>;
};

/**
 * The output of the Function run target.
 * The object contains the operations to apply to delivery options in checkout.
 */
export type CartDeliveryOptionsTransformRunResult = {
  /**
   * The ordered list of operations to apply to the list of
   * [delivery options](https://shopify.dev/docs/apps/build/checkout/delivery-shipping/delivery-options/build-function).
   */
  operations: Array<Operation>;
};
// ...
export type DeliveryCustomization = HasMetafields & {
  __typename?: 'DeliveryCustomization';
  /**
   * A [custom field](https://shopify.dev/docs/apps/build/custom-data) that stores additional information
   * about a Shopify resource, such as products, orders, and
   * [many more](https://shopify.dev/docs/api/admin-graphql/latest/enums/MetafieldOwnerType).
   * Using [metafields with Shopify Functions](https://shopify.dev/docs/apps/build/functions/input-output/metafields-for-input-queries)
   * enables you to customize the checkout experience.
   */
  metafield?: Maybe<Metafield>;
};


/** A customization representing how delivery options will be ordered, hidden, or renamed. */
export type DeliveryCustomizationMetafieldArgs = {
  key: Scalars['String']['input'];
  namespace?: InputMaybe<Scalars['String']['input']>;
};

/** List of different delivery method types. */
export enum DeliveryMethod {
  /** Local Delivery. */
  Local = 'LOCAL',
  /** None. */
  None = 'NONE',
  /** Shipping to a Pickup Point. */
  PickupPoint = 'PICKUP_POINT',
  /** Local Pickup. */
  PickUp = 'PICK_UP',
  /** Retail. */
  Retail = 'RETAIL',
  /** Shipping. */
  Shipping = 'SHIPPING'
}

/** An operation that hides a delivery option from a list that's offered to customers at checkout. */
export type DeliveryOptionHideOperation = {
  /** The handle of the delivery option to hide. */
  deliveryOptionHandle: Scalars['Handle']['input'];
};
// ...
export type Input = {
  __typename?: 'Input';
  /**
   * The cart where the Function is running. A cart contains the merchandise that a customer intends to purchase
   * and information about the customer, such as the customer's email address and phone number.
   */
  cart: Cart;
  /**
   * The backend logic that the Function is running to define how
   * [delivery options](https://shopify.dev/apps/build/checkout/delivery-shipping/delivery-options/build-function)
   * are sorted, hidden, or renamed. It includes the
   * [metafields](https://shopify.dev/docs/apps/build/custom-data)
   * that are associated with the customization.
   */
  deliveryCustomization: DeliveryCustomization;
  /**
   * The regional and language settings that determine how the Function
   * handles currency, numbers, dates, and other locale-specific values
   * during discount calculations. These settings are based on the store's configured
   * [localization practices](https://shopify.dev/docs/apps/build/functions/localization-practices-shopify-functions).
   */
  localization: Localization;
  /**
   * The exchange rate used to convert discounts between the shop's default
   * currency and the currency that displays to the customer during checkout.
   * For example, if a store operates in USD but a customer is viewing discounts in EUR,
   * then the presentment currency rate handles this conversion for accurate pricing.
   */
  presentmentCurrencyRate: Scalars['Decimal']['output'];
  /**
   * Information about the shop where the Function is running, including the shop's timezone
   * setting and associated [metafields](https://shopify.dev/docs/apps/build/custom-data).
   */
  shop: Shop;
};
// ...
export type Metafield = {
  __typename?: 'Metafield';
  /** The data that's stored in the metafield, using JSON format. */
  jsonValue: Scalars['JSON']['output'];
  /**
   * The [type of data](https://shopify.dev/apps/metafields/types) that the metafield stores in
   * the `value` field.
   */
  type: Scalars['String']['output'];
  /**
   * The data that's stored in the metafield. The data is always stored as a string,
   * regardless of the [metafield's type](https://shopify.dev/apps/metafields/types).
   */
  value: Scalars['String']['output'];
};
// ...
export type Operation =
  /** An operation that hides a delivery option from a list that's offered to customers at checkout. */
  { deliveryOptionHide: DeliveryOptionHideOperation; deliveryOptionMove?: never; deliveryOptionRename?: never; }
  |  /**
   * An operation that sorts a list of delivery options that are offered to customers at checkout.
   *
   * If you reorder shipping delivery options, then you are
   * [prohibited](https://shopify.dev/docs/apps/launch/app-requirements-checklist#prohibited-app-types).
   * from automatically selecting higher-priced delivery alternatives by default. The cheapest shipping delivery option
   * must always be the first option selected.
   */
  { deliveryOptionHide?: never; deliveryOptionMove: DeliveryOptionMoveOperation; deliveryOptionRename?: never; }
  |  /**
   * An operation that renames a delivery option that's offered to customers at checkout.
   *
   * The carrier name is automatically prepended to the delivery option title at checkout when using the
   * `DeliveryOptionRenameOperation` object, and can't be altered or omitted through the API. For example, if the carrier name
   * is **UPS** and the option is **Standard**, then you could change **UPS Standard** to **UPS Standard Shipping**,
   * but you couldn't change **UPS Standard** to **Standard Shipping**.
   */
  { deliveryOptionHide?: never; deliveryOptionMove?: never; deliveryOptionRename: DeliveryOptionRenameOperation; };
// ...
export type CartDeliveryOptionsTransformRunInput = { __typename?: 'Input', cart: { __typename?: 'Cart', deliveryGroups: Array<{ __typename?: 'CartDeliveryGroup', deliveryOptions: Array<{ __typename?: 'CartDeliveryOption', title?: string | null, handle: any, deliveryMethodType: DeliveryMethod }> }>, deliveryType?: { __typename?: 'Attribute', value?: string | null } | null, deliveryPromise?: { __typename?: 'Metafield', jsonValue: any } | null } };